feat: impossible to delete a nonexistent project

+++ src/modules/projects/usecases/deleteProject/DeleteProjectService.ts
@@ -1,3 +1,4 @@
+import { AppError } from "../../../../shared/errors/AppError";
 import { ProjectReposity } from "../../repositories/implementations/ProjectReposity";
 
 class DeleteProjectService {
@@ -5,6 +6,10 @@ class DeleteProjectService {
     constructor(private repository: ProjectReposity) { }
     async execute(id: string): Promise<void> {
         const project = await this.repository.findById(id);
+
+        if(!project) {
+            throw new AppError("Project not found", 404);
+        }
 
         await this.repository.deleteProject(project);
     }
